+++ R/tm_a_regression.R
@@ -147,8 +147,10 @@ tm_a_regression <- function(label = "Regression Analysis",
   # Important step, so we could easily consume it later
   if (is_ggplot2_args) ggplot2_args <- list(default = ggplot2_args)
 
-  check_slider_input(plot_height, allow_null = FALSE)
-  check_slider_input(plot_width)
+  checkmate::assert_numeric(plot_height, len = 3, any.missing = FALSE, finite = TRUE)
+  checkmate::assert_numeric(plot_height[c(2, 1, 3)], sorted = TRUE, .var.name = "plot_height")
+  checkmate::assert_numeric(plot_width, len = 3, any.missing = FALSE, null.ok = TRUE, finite = TRUE)
+  checkmate::assert_numeric(plot_width[c(2, 1, 3)], sorted = TRUE, null.ok = TRUE, .var.name = "plot_width")
 
   # Send ui args
   args <- as.list(environment())