+++ NEWS.md
@@ -16,6 +16,7 @@
 
 ## Miscellaneous
 * Refactored the defunct `teal.devel::data_extract_input` into its replacement `teal.devel::data_extract_ui`.
+* Replaced calls to `utils.nest::check_slider_input` with `checkmate::assert_numeric`.
 
 # teal.modules.general 0.2.11